Moved Camera Settings (getCameraSettings) to Configuration Class
